action: remove "latest" fallback to "stable" version
When the consumer specifies that they want the "latest" CUE version, we
must consult the Github API to find out what that version is. The API
sometimes (but rarely) fails to give us a list of CUE's releases. This
may be rate-limit related; it may be internal GitHub service errors -
we're not sure.

When experiencing that intermittent failure, the Action currently falls
back on a known-good CUE version (0.4.0) and installs it.

This is undesireable for a number of reasons, not least because it
requires us to keep the fallback version updated, here in code.

This commit changes that behaviour, and exposes the consumer of this
Action to these intermittent failures. This strategy was proposed and
agreed in https://github.com/cue-lang/cue/issues/2271.

Signed-off-by: Jonathan Matthews <github@hello.jonathanmatthews.com>