all: bump to use and support node20
* Upgrade 'vendored' GitHub action schema to use new node20 value.
* Use the latest version of actions/setup-node and actions/checkout in
  the workflows for this project to avoid triggering warnings about
  using node16-based actions.
* Upgrade the action that is this repo to be node20-based.
* Upgrade the version of CUE used in this repo to v0.7.1, current
  latest.
* Bump and fix various node dependencies to get things working again.
* Run 'npm ci' in workflows for this project to be more robust.

This change also disables the running of the jest tests as part of CI.
They are currently broken and myitcv does not have the experience to
debug and fix them. We still have coverage of the code however, because
the CI workflow for the project itself use the action to install CUE as
part of every commit.

Signed-off-by: Takuya Kosugiyama <re@itkq.jp>
Co-authored-by: Paul Jolly <paul@myitcv.io>